Move agent iterators to `stateful`
