+++ packages/engine/src/datastore/batch/mod.rs
@@ -1,11 +1,2 @@
 mod agent;
-pub mod boolean;
-pub mod iterators;
 pub mod migration;
-
-pub use stateful::{
-    agent::AgentBatch,
-    context::{AgentIndex, ContextBatch},
-    dataset::Dataset,
-    message::MessageBatch,
-};
