+++ packages/engine/src/datastore/mod.rs
@@ -28,15 +28,12 @@ pub mod tests {
 
     use ::arrow::array::{Array, BooleanBuilder, FixedSizeListBuilder};
     use rand::Rng;
+    use stateful::{agent, agent::AgentBatch};
 
     #[allow(clippy::wildcard_imports)] // Desigend as test-prelude
     use crate::datastore::test_utils::*;
     use crate::datastore::{
-        arrow::batch_conversion::IntoAgents,
-        batch::{iterators, AgentBatch},
-        error::Result,
-        table::state::State,
-        UUID_V4_LEN,
+        arrow::batch_conversion::IntoAgents, error::Result, table::state::State, UUID_V4_LEN,
     };
 
     #[test]
@@ -44,7 +41,7 @@ pub mod tests {
         pub fn modify_name(shmem_os_id: &str) -> Result<Vec<Option<String>>> {
             let mut state_batch = *AgentBatch::from_shmem_os_id(shmem_os_id)?;
             let record_batch = state_batch.batch.record_batch()?;
-            let mut column = iterators::record_batch::get_agent_name(record_batch)?;
+            let mut column = agent::arrow::record_batch::get_agent_name(record_batch)?;
             // `targets` values will be checked against shared memory data
             // in order to check if changes were flushed properly
             let mut targets = Vec::with_capacity(column.len());
@@ -70,7 +67,7 @@ pub mod tests {
                     }
                 }
             }
-            let change = iterators::record_batch::agent_name_as_array(record_batch, column)?;
+            let change = agent::arrow::record_batch::agent_name_as_array(record_batch, column)?;
             state_batch.batch.queue_change(change)?;
             state_batch.batch.flush_changes()?;
             Ok(targets)
@@ -116,7 +113,7 @@ pub mod tests {
         let batch_proxy: &AgentBatch = state_proxy.agent_pool().batch(0).unwrap();
         let record_batch = batch_proxy.batch.record_batch()?;
 
-        let names = iterators::record_batch::get_agent_name(record_batch)?;
+        let names = agent::arrow::record_batch::get_agent_name(record_batch)?;
         let agent_states = record_batch.into_agent_states(Some(&schema))?;
 
         targets.into_iter().enumerate().for_each(|(i, t)| match t {
