+++ packages/engine/src/simulation/package/context/packages/neighbors/map.rs
@@ -1,13 +1,11 @@
 use std::collections::HashSet;
 
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
+use stateful::context::AgentIndex;
 
-use crate::{
-    datastore::batch::AgentIndex,
-    simulation::{
-        package::context::packages::neighbors::{Result, TopologyConfig},
-        Error,
-    },
+use crate::simulation::{
+    package::context::packages::neighbors::{Result, TopologyConfig},
+    Error,
 };
 
 pub(super) type PositionSubType = f64;
