+++ packages/engine/src/simulation/package/output/packages/analysis/index_iter.rs
@@ -2,22 +2,19 @@ use std::cmp::Ordering;
 
 use arrow::datatypes::DataType;
 use float_cmp::approx_eq;
-use stateful::field::{FieldSpecMapAccessor, FieldTypeVariant};
+use stateful::{
+    agent,
+    field::{FieldSpecMapAccessor, FieldTypeVariant},
+};
 
-use crate::{
-    datastore::batch::iterators::agent::{
-        bool_iter, exists_iter, f64_iter, json_serialized_value_iter, json_value_iter_cols,
-        str_iter,
+use crate::simulation::package::output::packages::analysis::{
+    analyzer::{
+        AnalysisOperationRepr, ComparisonRepr, IndexIterator, OutputCreator, OutputRunner,
+        OutputRunnerCreator, ValueIterator, ValueIteratorCreator, ULPS,
     },
-    simulation::package::output::packages::analysis::{
-        analyzer::{
-            AnalysisOperationRepr, ComparisonRepr, IndexIterator, OutputCreator, OutputRunner,
-            OutputRunnerCreator, ValueIterator, ValueIteratorCreator, ULPS,
-        },
-        output::AnalysisSingleOutput,
-        value_iter::{value_iterator_filter, value_iterator_mapper},
-        Error, Result,
-    },
+    output::AnalysisSingleOutput,
+    value_iter::{value_iterator_filter, value_iterator_mapper},
+    Error, Result,
 };
 
 fn index_iterator_f64_filter(
@@ -701,7 +698,7 @@ fn default_first_getter(
 
     let first_field = first_field.to_string();
     let a: ValueIteratorCreator = Box::new(move |agents: &_| {
-        let iterator = json_value_iter_cols(agents, &first_field, &data_type)?;
+        let iterator = agent::arrow::json_value_iter_cols(agents, &first_field, &data_type)?;
         Ok(iterator as ValueIterator<'_>)
     });
     Ok(a)
@@ -748,7 +745,7 @@ pub(super) fn index_iterator_mapper_creator(
         }
         FieldTypeVariant::AnyType => {
             let a: ValueIteratorCreator = Box::new(move |agents| {
-                let iterator = json_serialized_value_iter(agents, &first_field)?;
+                let iterator = agent::arrow::json_serialized_value_iter(agents, &first_field)?;
                 Ok(Box::new(iterator) as ValueIterator<'_>)
             });
             a
