+++ packages/engine/src/worker/runner/javascript/mod.rs
@@ -26,7 +26,10 @@ use memory::{
     arrow::{ArrowBatch, ColumnChange},
     shared_memory::{arrow_continuation, Metaversion, Segment},
 };
-use stateful::{field::PackageId, globals::Globals, state::StateWriteProxy};
+use stateful::{
+    agent::AgentBatch, field::PackageId, globals::Globals, message::MessageBatch,
+    state::StateWriteProxy,
+};
 use tokio::{
     sync::mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender},
     task::JoinError,
@@ -36,7 +39,6 @@ use tracing::{Instrument, Span};
 pub use self::error::{Error, Result};
 use crate::{
     datastore::{
-        batch::{AgentBatch, MessageBatch},
         shared_store::SharedStore,
         table::{
             sync::{ContextBatchSync, StateSync, WaitableStateSync},
